pyproject.toml
[project]
name = "chat-bots-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-telegram-bot==13.15",
    "redis>=6.2.0",
    "vk-api>=11.9.9",
]
